requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
python-dotenv
sqlalchemy
aiosqlite
pytest
google-generativeai
